libc: add some stub functions

+++ servers/mm/proto.h
@@ -30,9 +30,9 @@ int free_mem(phys_bytes base, phys_bytes len);
 phys_bytes alloc_pages(int nr_pages, int memflags);
 
 void vmem_init(void* mem_start, size_t free_mem_size);
-void* alloc_vmem(phys_bytes* phys_addr, int memsize, int reason);
+void* alloc_vmem(phys_bytes* phys_addr, size_t memsize, int reason);
 void* alloc_vmpages(int nr_pages);
-void free_vmem(void* base, int len);
+void free_vmem(void* base, size_t len);
 void free_vmpages(void* base, int nr_pages);
 
 void slabs_init();
