+++ servers/mm/vmalloc.c
@@ -82,7 +82,7 @@ void vmem_init(void* mem_start, size_t free_mem_size)
  *
  * @return  The base of the memory just allocated.
  *****************************************************************************/
-void* alloc_vmem(phys_bytes* phys_addr, int memsize, int reason)
+void* alloc_vmem(phys_bytes* phys_addr, size_t memsize, int reason)
 {
     /* avoid recursive allocation */
     static int level = 0;
@@ -238,7 +238,7 @@ void free_vmpages(void* base, int nr_pages)
  *
  * @return  Zero if success.
  *****************************************************************************/
-void free_vmem(void* base, int len)
+void free_vmem(void* base, size_t len)
 {
     if (!pt_init_done) return;
